Fix the menu event and initializers, remove menu test
