+++ src/main/java/io/github/fabriccommunity/everything/impl/function/ClientInitRunner.java
@@ -1,4 +1,4 @@
-package io.github.fabriccommunity.everythingtest;
+package io.github.fabriccommunity.everything.impl.function;
 
 import io.github.fabriccommunity.everything.api.functional.FunctionalModInitializer;
 import io.github.fabriccommunity.everything.api.functional.IO;
@@ -7,9 +7,9 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 
 @Environment(EnvType.CLIENT)
-public final class ExampleModClient implements ClientModInitializer {
+public final class ClientInitRunner implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
-        IO.executeUnsafe(ExampleMod.runInitializers("everything-api/functional/client", FunctionalModInitializer.ClientOnly::onInitializeClient, FunctionalModInitializer.ClientOnly.class));
+        IO.executeUnsafe(CommonInitRunner.runInitializers("everything-api/functional/client", FunctionalModInitializer.ClientOnly::onInitializeClient, FunctionalModInitializer.ClientOnly.class));
     }
 }
