+++ src/main/java/io/github/fabriccommunity/everything/impl/function/ServerInitRunner.java
@@ -1,4 +1,4 @@
-package io.github.fabriccommunity.everythingtest;
+package io.github.fabriccommunity.everything.impl.function;
 
 import io.github.fabriccommunity.everything.api.functional.FunctionalModInitializer;
 import io.github.fabriccommunity.everything.api.functional.IO;
@@ -7,9 +7,9 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 
 @Environment(EnvType.SERVER)
-public final class ExampleModServer implements DedicatedServerModInitializer {
+public final class ServerInitRunner implements DedicatedServerModInitializer {
     @Override
     public void onInitializeServer() {
-        IO.executeUnsafe(ExampleMod.runInitializers("everything-api/functional/server", FunctionalModInitializer.ServerOnly::onInitializeServer, FunctionalModInitializer.ServerOnly.class));
+        IO.executeUnsafe(CommonInitRunner.runInitializers("everything-api/functional/server", FunctionalModInitializer.ServerOnly::onInitializeServer, FunctionalModInitializer.ServerOnly.class));
     }
 }
